# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VisaVantage API – marketplace for students and employers
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs with optional filters
      parameters:
        - name: visaType
          in: query
          schema:
            type: string
            enum: [STAMP_2, STAMP_1G, GRADUATE_VISA, STUDENT_VISA, ANY]
        - name: jobType
          in: query
          schema:
            type: string
            enum: [PART_TIME, FULL_TIME, INTERNSHIP, FREELANCE]
        - name: location
          in: query
          schema:
            type: string
        - name: minPay
          in: query
          schema:
            type: number
        - name: maxPay
          in: query
          schema:
            type: number
        - name: category
          in: query
          schema:
            type: string
        - name: isRemote
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"

    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job posting (employer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job with visa eligibility for logged-in student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job (employer only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job (employer only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── APPLICATIONS ──────────────────────────────────────────────────
...
  /cv:
    get:
      operationId: getCv
      tags: [cv]
      summary: Get student's saved CV data
      responses:
        "200":
          description: CV data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CvData"
        "404":
          description: No CV yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: saveCv
      tags: [cv]
      summary: Save/update structured CV data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CvInput"
      responses:
        "200":
          description: Saved CV data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CvData"
  /cv/score:
    get:
      operationId: scoreCv
      tags: [cv]
      summary: Score student CV against a job description
      parameters:
        - name: jobId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CvScoreResult"

  # ── NOTIFICATIONS ─────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH SCHEMAS ────────────────────────────────────────────────
...
    Job:
      type: object
      required: [id, employerId, title, jobType, status, createdAt]
      properties:
        id:
          type: integer
        employerId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        jobType:
          type: string
          enum: [PART_TIME, FULL_TIME, INTERNSHIP, FREELANCE]
        visaEligible:
          type: array
          items:
            type: string
        hoursPerWeek:
          type: ["number", "null"]
        payRate:
          type: ["number", "null"]
        payPeriod:
          type: ["string", "null"]
          enum: [HOURLY, WEEKLY, MONTHLY, null]
        location:
          type: ["string", "null"]
        isRemote:
          type: boolean
        isFeatured:
          type: boolean
        status:
          type: string
          enum: [ACTIVE, CLOSED, DRAFT]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        employer:
          $ref: "#/components/schemas/EmployerSummary"
    JobDetail:
      type: object
      required: [id, employerId, title, jobType, status, createdAt]
      properties:
        id:
          type: integer
        employerId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        jobType:
          type: string
          enum: [PART_TIME, FULL_TIME, INTERNSHIP, FREELANCE]
        visaEligible:
          type: array
          items:
            type: string
        hoursPerWeek:
          type: ["number", "null"]
        payRate:
          type: ["number", "null"]
        payPeriod:
          type: ["string", "null"]
          enum: [HOURLY, WEEKLY, MONTHLY, null]
        location:
          type: ["string", "null"]
        isRemote:
          type: boolean
        isFeatured:
          type: boolean
        status:
          type: string
          enum: [ACTIVE, CLOSED, DRAFT]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        employer:
          $ref: "#/components/schemas/EmployerSummary"
        visaEligibilityStatus:
          type: ["string", "null"]
          enum: [ELIGIBLE, CHECK_HOURS, NOT_ELIGIBLE, null]
        applicationCount:
          type: integer
        hasApplied:
          type: boolean
    EmployerSummary:
      type: object
      required: [id, companyName]
      properties:
        id:
          type: integer
        companyName:
          type: string
        logoUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        trustScore:
          type: ["number", "null"]
        sector:
          type: ["string", "null"]
    JobListResponse:
      type: object
      required: [jobs, total, page, limit]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    JobInput:
      type: object
      required: [title, jobType]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        jobType:
          type: string
          enum: [PART_TIME, FULL_TIME, INTERNSHIP, FREELANCE]
        visaEligible:
          type: array
          items:
            type: string
        hoursPerWeek:
          type: number
        payRate:
          type: number
        payPeriod:
          type: string
          enum: [HOURLY, WEEKLY, MONTHLY]
        location:
          type: string
        isRemote:
          type: boolean
        isFeatured:
          type: boolean
        expiresAt:
          type: string
          format: date-time
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        jobType:
          type: string
          enum: [PART_TIME, FULL_TIME, INTERNSHIP, FREELANCE]
        visaEligible:
          type: array
          items:
            type: string
        hoursPerWeek:
          type: number
        payRate:
          type: number
        payPeriod:
          type: string
          enum: [HOURLY, WEEKLY, MONTHLY]
        location:
          type: string
        isRemote:
          type: boolean
        isFeatured:
          type: boolean
        status:
          type: string
          enum: [ACTIVE, CLOSED, DRAFT]
        expiresAt:
          type: string
          format: date-time

    # ── APPLICATION SCHEMAS ─────────────────────────────────────────
...
    CvData:
      type: object
      required: [id, studentId, updatedAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        parsedJson:
          type: object
          additionalProperties: true
        aiScore:
          type: ["number", "null"]
        updatedAt:
          type: string
          format: date-time
    CvInput:
      type: object
      required: [parsedJson]
      properties:
        parsedJson:
          type: object
          additionalProperties: true
    CvScoreResult:
      type: object
      required: [score, tips]
      properties:
        score:
          type: number
        tips:
          type: array
          items:
            type: string
        strengths:
          type: array
          items:
            type: string

    # ── NOTIFICATION SCHEMAS ────────────────────────────────────────
...