# lib/api-zod/src/generated/types/jobDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VisaVantage API – marketplace for students and employers
 * OpenAPI spec version: 0.1.0
 */
import type { EmployerSummary } from './employerSummary';
import type { JobDetailJobType } from './jobDetailJobType';
import type { JobDetailPayPeriod } from './jobDetailPayPeriod';
import type { JobDetailStatus } from './jobDetailStatus';
import type { JobDetailVisaEligibilityStatus } from './jobDetailVisaEligibilityStatus';

export interface JobDetail {
  id: number;
  employerId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  jobType: JobDetailJobType;
  visaEligible?: string[];
  /** @nullable */
  hoursPerWeek?: number | null;
  /** @nullable */
  payRate?: number | null;
  /** @nullable */
  payPeriod?: JobDetailPayPeriod;
  /** @nullable */
  location?: string | null;
  isRemote?: boolean;
  isFeatured?: boolean;
  status: JobDetailStatus;
  /** @nullable */
  expiresAt?: Date | null;
  createdAt: Date;
  employer?: EmployerSummary;
  /** @nullable */
  visaEligibilityStatus?: JobDetailVisaEligibilityStatus;
  applicationCount?: number;
  hasApplied?: boolean;
}